+++ R/db_ops.R
@@ -77,26 +77,35 @@ rq_execute <- function(db, q) {
 #'
 #' @param db Connection handle
 #' @param table_name character table name
+#' @param ... not used, force later argument to bind by name
+#' @param qualifiers optional named ordered vector of strings carrying additional db hierarchy terms, such as schema.
 #' @return logical TRUE if table exists.
 #'
 #' @seealso \code{\link{db_td}}
 #'
 #' @export
 #'
-rq_table_exists <- function(db, table_name) {
+rq_table_exists <- function(db, table_name,
+                            ...,
+                            qualifiers = NULL) {
   if(is.null(db)) {
     stop("rquery::rq_table_exists db was null")
   }
+  wrapr::stop_if_dot_args(substitute(list(...)),
+                          "rquery::rq_table_exists")
   # first shot- see if it is a db info with function overrriden
   connection <- db
   connection_options <- NULL
   if("rquery_db_info" %in% class(db)) {
     f <- db$rq_table_exists
     if(!is.null(f)) {
-      return(f(db, table_name))
+      return(f(db, table_name, qualifiers = qualifiers))
     }
     connection_options <- db$connection_options
     connection <- db$connection
+    q_table_name <- db$quote_table_name(db, table_name, qualifiers = qualifiers)
+  } else {
+    q_table_name <- quote_table_name(db, table_name, qualifiers = qualifiers)
   }
   if(is.null(connection)) {
     stop("rquery::rq_table_exists db$connection was null")
@@ -106,7 +115,7 @@ rq_table_exists <- function(db, table_name) {
     return(DBI::dbExistsTable(connection, table_name))
   }
   q <- paste0("SELECT * FROM ",
-              quote_identifier(db, table_name),
+              q_table_name,
               " LIMIT 1")
   # others do a 0=1 thing, may be faster but let's try this.
   tryCatch(
@@ -154,6 +163,9 @@ rq_colnames <- function(db, table_name,
     }
     connection_options <- db$connection_options
     connection <- db$connection
+    q_table_name <- db$quote_table_name(db, table_name, qualifiers = qualifiers)
+  } else {
+    q_table_name <- quote_table_name(db, table_name, qualifiers = qualifiers)
   }
   if(is.null(connection)) {
     stop("rquery::rq_colnames db$connection was null")
@@ -161,11 +173,12 @@ rq_colnames <- function(db, table_name,
   # DBI::dbListFields fails intermitnently, and sometimes gives wrong results
   # filed as: https://github.com/tidyverse/dplyr/issues/3204
   if(getDBOption(db, "use_DBI_dbListFields", FALSE, connection_options) && requireNamespace("DBI", quietly = TRUE)) {
+    # this path doesn't handle schemas
     return(DBI::dbListFields(connection, table_name))
   }
-  # below is going to have issues to to R-column name conversion!
+  # below is going to have issues to R-column name conversion!
   q <- paste0("SELECT * FROM ",
-              quote_identifier(db, table_name),
+              q_table_name,
               " LIMIT 1")
   v <- rq_get_query(db, q)
   colnames(v)
@@ -180,6 +193,7 @@ rq_colnames <- function(db, table_name,
 #' @param db Connection handle.
 #' @param table_name character table name referring to a non-empty table.
 #' @param ... force later arguments to bind by name.
+#' @param qualifiers optional named ordered vector of strings carrying additional db hierarchy terms, such as schema.
 #' @param prefer_not_NA logical, if TRUE try to find an non-NA example for all columns (FALSE just for logical columns).
 #' @param force_check logical, if TRUE perform checks regardless of check_logical_column_types option setting.
 #' @return single row data.frame with example values, not all values necessarily from same database row.
@@ -224,9 +238,10 @@ rq_colnames <- function(db, table_name,
 #' @export
 #'
 rq_coltypes <- function(db, table_name,
-                         ...,
-                         prefer_not_NA = FALSE,
-                         force_check = FALSE) {
+                        ...,
+                        qualifiers = NULL,
+                        prefer_not_NA = FALSE,
+                        force_check = FALSE) {
   if(is.null(db)) {
     stop("rquery::rq_coltypes db was null")
   }
@@ -239,11 +254,15 @@ rq_coltypes <- function(db, table_name,
     f <- db$rq_coltypes
     if(!is.null(f)) {
       return(f(db, table_name,
+               qualifiers = qualifiers,
                prefer_not_NA = prefer_not_NA,
                force_check = force_check))
     }
     connection_options <- db$connection_options
     connection <- db$connection
+    q_table_name <- db$quote_table_name(db, table_name, qualifiers = qualifiers)
+  } else {
+    q_table_name <- quote_table_name(db, table_name, qualifiers = qualifiers)
   }
   if(is.null(connection)) {
     stop("rquery::rq_coltypes db$connection was null")
@@ -251,8 +270,7 @@ rq_coltypes <- function(db, table_name,
   # RSQLite returns logical type for any returned column
   # that is entirely NA, regardless of storage type.
   # below is going to have issues to to R-column name conversion!
-  tn <- quote_identifier(db, table_name)
-  q <- paste("SELECT * FROM", tn, "LIMIT 1")
+  q <- paste("SELECT * FROM", q_table_name, "LIMIT 1")
   v <- rq_get_query(db, q)
   if((nrow(v)>0) &&
      (force_check || getDBOption(db, "check_logical_column_types", FALSE, connection_options))) {
@@ -261,7 +279,7 @@ rq_coltypes <- function(db, table_name,
       if(is.na(cv)) {
         if(prefer_not_NA || is.logical(cv)) {
           cn <- quote_identifier(db, ci)
-          qi <- paste("SELECT", cn, "FROM ", tn, "WHERE", cn, "IS NOT NULL LIMIT 1")
+          qi <- paste("SELECT", cn, "FROM ", q_table_name, "WHERE", cn, "IS NOT NULL LIMIT 1")
           vi <- rq_get_query(db, qi)
           if(nrow(vi)>0) {
             v[[ci]] <- vi[[ci]]
@@ -278,16 +296,22 @@ rq_coltypes <- function(db, table_name,
 #'
 #' @param db database connection.
 #' @param table_name character, name of table to create.
+#' @param ... not used, force later argument to bind by name
+#' @param qualifiers optional named ordered vector of strings carrying additional db hierarchy terms, such as schema.
 #' @return logical TRUE if table existed, else FALSE
 #'
 #' @seealso \code{\link{db_td}}
 #'
 #' @export
 #'
-rq_remove_table <- function(db, table_name) {
+rq_remove_table <- function(db, table_name,
+                            ...,
+                            qualifiers = NULL) {
   if(is.null(db)) {
     stop("rquery::rq_remove_table db was null")
   }
+  wrapr::stop_if_dot_args(substitute(list(...)),
+                          "rquery::rq_remove_table")
   # first shot- see if it is a db info with function overrriden
   connection_options <- NULL
   connection <- db
@@ -298,18 +322,21 @@ rq_remove_table <- function(db, table_name) {
     }
     connection_options <- db$connection_options
     connection <- db$connection
+    q_table_name <- db$quote_table_name(db, table_name, qualifiers = qualifiers)
+  } else {
+    q_table_name <- quote_table_name(db, table_name, qualifiers = qualifiers)
   }
   if(is.null(connection)) {
     stop("rquery::rq_remove_table db$connection was null")
   }
   if(!is.null(table_name)) {
-    if(rq_table_exists(db, table_name)) {
+    if(rq_table_exists(db, table_name, qualifiers = qualifiers)) {
       if(getDBOption(db, "use_DBI_dbRemoveTable", FALSE, connection_options) && requireNamespace("DBI", quietly = TRUE)) {
         DBI::dbRemoveTable(connection, table_name)
       } else {
         rq_execute(db,
-                    paste("DROP TABLE",
-                           quote_identifier(db, table_name)))
+                   paste("DROP TABLE",
+                         q_table_name))
       }
       return(TRUE)
     }
@@ -327,8 +354,18 @@ connection_is_sparklyr <- function(db) {
     db <- db$connection
   }
   length(intersect(c("spark_connection", "spark_shell_connection"),
-                             class(db)))>=1
+                   class(db)))>=1
 }
+
+maybe_dbi_table_id <- function(table_name, qualifiers) {
+  # https://github.com/r-dbi/odbc/issues/91
+  if("schema" %in% names(qualifiers)) {
+    return(DBI::Id(table = table_name, schema = qualifiers[["schema"]]))
+  }
+  # DBI::Id(table = table_name) # RPostgreSQL can't use this form in some cases
+  table_name
+}
+
 
 #' Copy local R table to remote data handle.
 #'
@@ -336,6 +373,7 @@ connection_is_sparklyr <- function(db) {
 #' @param table_name name of table to create.
 #' @param d data.frame to copy to database.
 #' @param ... force later argument to be by name
+#' @param qualifiers optional named ordered vector of strings carrying additional db hierarchy terms, such as schema.
 #' @param overwrite logical, if TRUE try to overwrite existing table.
 #' @param temporary logical, if TRUE try to mark table as temporary.
 #' @param rowidcolumn character, name to land row-ids.
@@ -359,10 +397,11 @@ connection_is_sparklyr <- function(db) {
 #' @export
 #'
 rq_copy_to <- function(db, table_name, d,
-                        ...,
-                        overwrite = FALSE,
-                        temporary = TRUE,
-                        rowidcolumn = NULL) {
+                       ...,
+                       qualifiers = NULL,
+                       overwrite = FALSE,
+                       temporary = TRUE,
+                       rowidcolumn = NULL) {
   if(is.null(db)) {
     stop("rquery::rq_copy_to db was null")
   }
@@ -375,6 +414,7 @@ rq_copy_to <- function(db, table_name, d,
     f <- db$rq_copy_to
     if(!is.null(f)) {
       return(f(db, table_name, d,
+               qualifiers = qualifiers,
                overwrite = overwrite,
                temporary = temporary,
                rowidcolumn = rowidcolumn))
@@ -411,24 +451,24 @@ rq_copy_to <- function(db, table_name, d,
   if(!requireNamespace("DBI", quietly = TRUE)) {
     stop("rquery::rq_copy_to without per-connection implemention need DBI package")
   }
-  if(rq_table_exists(db, table_name)) {
+  if(rq_table_exists(db, table_name, qualifiers = qualifiers)) {
     if(overwrite) {
       # sparklyr 0.7.0 can't take overwrite argument
-      rq_remove_table(db, table_name)
+      rq_remove_table(db, table_name, qualifiers = qualifiers)
     } else {
       stop(paste("rquery::rq_copy_to table", table_name, "exists and overwrite==FALSE"))
     }
   }
   if(can_set_temp) {
-     if(can_set_rownames) {
+    if(can_set_rownames) {
       DBI::dbWriteTable(connection,
-                        table_name,
+                        maybe_dbi_table_id(table_name, qualifiers = qualifiers),
                         d,
                         temporary = temporary,
                         row.names = FALSE)
     } else {
       DBI::dbWriteTable(connection,
-                        table_name,
+                        maybe_dbi_table_id(table_name, qualifiers = qualifiers),
                         d,
                         temporary = temporary)
     }
@@ -438,44 +478,52 @@ rq_copy_to <- function(db, table_name, d,
     }
     if(can_set_rownames) {
       DBI::dbWriteTable(connection,
-                        table_name,
+                        maybe_dbi_table_id(table_name, qualifiers = qualifiers),
                         d,
                         row.names = FALSE)
     } else {
       DBI::dbWriteTable(connection,
-                        table_name,
+                        maybe_dbi_table_id(table_name, qualifiers = qualifiers),
                         d)
     }
   }
-  db_td(db, table_name)
+  db_td(db, table_name, qualifiers = qualifiers)
 }
 
 #' Count rows and return as numeric
 #'
 #' @param db database connection
 #' @param table_name character, name of table
+#' @param ... not used, force later argument to bind by name
+#' @param qualifiers optional named ordered vector of strings carrying additional db hierarchy terms, such as schema.
 #' @return numeric row count
 #'
 #' @seealso \code{\link{db_td}}
 #'
 #' @export
 #'
-rq_nrow <- function(db, table_name) {
+rq_nrow <- function(db, table_name,
+                    ...,
+                    qualifiers = NULL) {
   if(is.null(db)) {
     stop("rquery::rq_nrow db was null")
   }
+  wrapr::stop_if_dot_args(substitute(list(...)),
+                          "rquery::rq_nrow")
   # first shot- see if it is a db info with function overrriden
   if("rquery_db_info" %in% class(db)) {
     f <- db$rq_nrow
     if(!is.null(f)) {
-      return(f(db, table_name))
+      return(f(db, table_name, qualifiers = qualifiers))
     }
+    q_table_name <- db$quote_table_name(db, table_name, qualifiers = qualifiers)
+  } else {
+    q_table_name <- quote_table_name(db, table_name, qualifiers = qualifiers)
   }
   nrowst <- rq_get_query(
     db,
     paste0("SELECT COUNT(1) FROM ",
-           quote_identifier(db,
-                                  table_name)))
+           q_table_name))
   # integer64 was coming back from RPostgres
   # and that does not work as numeric in pmin()
   nrows <- as.numeric(nrowst[[1]][[1]])
@@ -558,17 +606,27 @@ rq_connection_advice <- function(db) {
     opts[[paste(c("rquery", cname, "use_DBI_dbRemoveTable"), collapse = ".")]] <- FALSE
     opts[[paste(c("rquery", cname, "use_DBI_dbExistsTable"), collapse = ".")]] <- FALSE # fails on some CREATE AS tables
   }
+  if(cname =="PqConnection") { # RPostgres::Postgres()
+    # schema issues
+    opts[[paste(c("rquery", cname, "use_DBI_dbListFields"), collapse = ".")]] <- FALSE
+    opts[[paste(c("rquery", cname, "use_DBI_dbRemoveTable"), collapse = ".")]] <- FALSE
+    opts[[paste(c("rquery", cname, "use_DBI_dbExistsTable"), collapse = ".")]] <- FALSE
+  }
   opts
 }
 
 
-brute_rm_table <- function(db, table_name) {
+brute_rm_table <- function(db, table_name,
+                           ...,
+                           qualifiers = NULL) {
+  wrapr::stop_if_dot_args(substitute(list(...)),
+                          "rquery:::brute_rm_table")
   if(is.null(db)) {
     stop("rquery::brute_rm_table db was null")
   }
   tryCatch(
     rq_execute(db, paste("DROP TABLE",
-                         quote_identifier(db, table_name))),
+                         quote_table_name(db, table_name, qualifiers = qualifiers))),
     error = function(e) {e},
     warning = function(w) {w})
   NULL
@@ -578,6 +636,7 @@ brute_rm_table <- function(db, table_name) {
 #'
 #' These settings are estimated by experiments.  This is not
 #' the full set of options- but just the ones tested here.
+#' Note: tests are currently run in the default schema.
 #'
 #' @param db database connection handle.
 #' @param ... force later arguments to bind by name.
@@ -602,9 +661,9 @@ brute_rm_table <- function(db, table_name) {
 #' @export
 #'
 rq_connection_tests <- function(db,
-                                 ...,
-                                 overrides = NULL,
-                                 use_advice = TRUE) {
+                                ...,
+                                overrides = NULL,
+                                use_advice = TRUE) {
   if(is.null(db)) {
     stop("rquery::rq_connection_tests db was null")
   }
@@ -698,7 +757,7 @@ rq_connection_tests <- function(db,
   tryCatch(
     {
       DBI::dbExecute(connection, paste("DROP TABLE",
-                               obscure_name_q))
+                                       obscure_name_q))
       opts[[paste(c("rquery", cname, "use_DBI_dbExecute"), collapse = ".")]] <- TRUE
     },
     error = function(e) { e },
@@ -708,22 +767,23 @@ rq_connection_tests <- function(db,
   tryCatch(
     {
       DBI::dbGetQuery(connection, paste("CREATE TEMPORARY TABLE",
-                               obscure_name_q,
-                               "( x INT )"))
+                                        obscure_name_q,
+                                        "( x INT )"))
       opts[[paste(c("rquery", cname, "create_temporary"), collapse = ".")]] <- TRUE
     },
     error = function(e) { e },
     warning = function(w) { w })
-  # check on temporary view
-  tryCatch(
-    {
-      DBI::dbGetQuery(connection, paste("CREATE TEMPORARY VIEW",
-                                        obscure_name_q,
-                                        "( x INT )"))
-      opts[[paste(c("rquery", cname, "control_temporary_view"), collapse = ".")]] <- TRUE
-    },
-    error = function(e) { e },
-    warning = function(w) { w })
+  # # check on temporary view NOT CORRECT, NEED 2nd table name
+  # tryCatch(
+  #   {
+  #     DBI::dbGetQuery(connection, paste("CREATE TEMPORARY VIEW",
+  #                                       obscure_name_q,
+  #                                       "AS SELECT * FROM ",
+  #                                       obscure_name))
+  #     opts[[paste(c("rquery", cname, "control_temporary_view"), collapse = ".")]] <- TRUE
+  #   },
+  #   error = function(e) { e },
+  #   warning = function(w) { w })
   brute_rm_table(db, obscure_name)
   # see if NA columns masquerade as logical
   # (RSQLite has this property for some derived columns)
@@ -734,8 +794,8 @@ rq_connection_tests <- function(db,
                   want = c(1, 0),
                   stringsAsFactors=FALSE)
   d <- rq_copy_to(db, obscure_name, d,
-              overwrite = TRUE,
-              temporary = TRUE)
+                  overwrite = TRUE,
+                  temporary = TRUE)
   # make column refs not look like unbound references
   w <- NULL # don't appear unbound
   want <- NULL # don't appear unbound
@@ -749,14 +809,14 @@ rq_connection_tests <- function(db,
   yn <- NULL # don't appear unbound
   local_sample <- d %.>%
     extend(.,
-               wc %:=% ifelse(w>1, "x", "y"),
-               wn %:=% ifelse(w>1, 1, 2),
-               xc %:=% ifelse(x>1, "x", "y"),
-               xn %:=% ifelse(x>1, 1, 2),
-               yc %:=% ifelse(y=="a", "x", "y"),
-               yn %:=% ifelse(y=="a", "x", "y")) %.>%
+           wc %:=% ifelse(w>1, "x", "y"),
+           wn %:=% ifelse(w>1, 1, 2),
+           xc %:=% ifelse(x>1, "x", "y"),
+           xn %:=% ifelse(x>1, 1, 2),
+           yc %:=% ifelse(y=="a", "x", "y"),
+           yn %:=% ifelse(y=="a", "x", "y")) %.>%
     select_rows(.,
-                    want == 1) %.>%
+                want == 1) %.>%
     execute(db, .)
   logical_col <- vapply(colnames(local_sample),
                         function(ci) is.logical(local_sample[[ci]]), logical(1))