+++ NEWS.md
@@ -1,4 +1,10 @@
 
+# rquery 1.3.4 2019/06/14
+
+ * Work on schema qualifications.
+ * Allow empty project.
+ * Better function re-mapping.
+ 
 # rquery 1.3.3 2019/06/01
 
  * Clean up test_set_indicator() error message.